[Feat/#230] 배차 알고리즘 실행 트리거 구현

+++ backend/src/main/java/hyfive/gachita/dispatch/dto/CarScheduleDto.java
@@ -4,10 +4,10 @@ import java.time.LocalTime;
 
 public record CarScheduleDto(
         Long centerId,
-        Long centerLat, // 제거?
-        Long centerLng, // 제거?
+        double centerLat,
+        double centerLng,
         Long carId,
-        Long carCapacity,
+        int carCapacity,
         Long rentalId,
         LocalTime rentalStartTime,
         LocalTime rentalEndTime
